pyproject.toml
[project]
name = "flights"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bs4>=0.0.2",
    "fast-flights>=2.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
